+++ src/AIServer/ServerDlg.cpp
@@ -200,7 +200,7 @@ bool CServerDlg::CreateNpcThread()
 		}
 	}
 
-	printf("[Monster Init - %d, threads=%lld]\n", m_TotalNPC, (long long) m_arNpcThread.size());
+	printf("[Monster Init - %d, threads=%lld]\n", (uint16) m_TotalNPC, (long long) m_arNpcThread.size());
 	return true;
 }
